Convert shortcut services from .svelte.ts to just .ts
Because $effects are difficult to work with in unit tests.
