+++ apps/desktop/src/components/GlobalSettingsMenuAction.svelte
@@ -1,11 +1,13 @@
 <script lang="ts">
 	import { goto } from '$app/navigation';
 	import { newSettingsPath } from '$lib/routes/routes.svelte';
-	import { SHORTCUT_SERVICE } from '$lib/shortcuts/shortcutService.svelte';
+	import { SHORTCUT_SERVICE } from '$lib/shortcuts/shortcutService';
 	import { inject } from '@gitbutler/shared/context';
 
 	const shortcutService = inject(SHORTCUT_SERVICE);
-	shortcutService.on('global-settings', () => {
-		goto(newSettingsPath());
-	});
+	$effect(() =>
+		shortcutService.on('global-settings', () => {
+			goto(newSettingsPath());
+		})
+	);
 </script>
