+++ apps/desktop/src/components/ProjectSettingsMenuAction.svelte
@@ -5,12 +5,11 @@
 	import { PROJECTS_SERVICE } from '$lib/project/projectsService';
 	import { projectSettingsPath } from '$lib/routes/routes.svelte';
 	import { SETTINGS } from '$lib/settings/userSettings';
-	import { SHORTCUT_SERVICE } from '$lib/shortcuts/shortcutService.svelte';
+	import { SHORTCUT_SERVICE } from '$lib/shortcuts/shortcutService';
 	import * as events from '$lib/utils/events';
-	import { unsubscribe } from '$lib/utils/unsubscribe';
 	import { getEditorUri, openExternalUrl, showFileInFolder } from '$lib/utils/url';
 	import { inject } from '@gitbutler/shared/context';
-	import { onMount } from 'svelte';
+	import { mergeUnlisten } from '@gitbutler/ui/utils/mergeUnlisten';
 
 	const { projectId }: { projectId: string } = $props();
 
@@ -19,46 +18,38 @@
 	const userSettings = inject(SETTINGS);
 	const shortcutService = inject(SHORTCUT_SERVICE);
 
-	shortcutService.on('project-settings', () => {
-		goto(projectSettingsPath(projectId));
-	});
-
-	shortcutService.on('open-in-vscode', async () => {
-		const project = await projectsService.fetchProject(projectId);
-		if (!project) {
-			throw new Error(`Project not found: ${projectId}`);
-		}
-		openExternalUrl(
-			getEditorUri({
-				schemeId: $userSettings.defaultCodeEditor.schemeIdentifer,
-				path: [vscodePath(project.path)],
-				searchParams: { windowId: '_blank' }
+	$effect(() =>
+		mergeUnlisten(
+			shortcutService.on('project-settings', () => {
+				goto(projectSettingsPath(projectId));
+			}),
+			shortcutService.on('history', () => {
+				$showHistoryView = !$showHistoryView;
+			}),
+			shortcutService.on('open-in-vscode', async () => {
+				const project = await projectsService.fetchProject(projectId);
+				if (!project) {
+					throw new Error(`Project not found: ${projectId}`);
+				}
+				openExternalUrl(
+					getEditorUri({
+						schemeId: $userSettings.defaultCodeEditor.schemeIdentifer,
+						path: [vscodePath(project.path)],
+						searchParams: { windowId: '_blank' }
+					})
+				);
+			}),
+			events.on('openHistory', () => {
+				$showHistoryView = true;
+			}),
+			shortcutService.on('show-in-finder', async () => {
+				const project = await projectsService.fetchProject(projectId);
+				if (!project) {
+					throw new Error(`Project not found: ${projectId}`);
+				}
+				// Show the project directory in the default file manager (cross-platform)
+				await showFileInFolder(project.path);
 			})
-		);
-	});
-
-	shortcutService.on('show-in-finder', async () => {
-		const project = await projectsService.fetchProject(projectId);
-		if (!project) {
-			throw new Error(`Project not found: ${projectId}`);
-		}
-		// Show the project directory in the default file manager (cross-platform)
-		await showFileInFolder(project.path);
-	});
-
-	shortcutService.on('history', () => {
-		$showHistoryView = !$showHistoryView;
-	});
-
-	const unsubscribeHistoryButton = unsubscribe(
-		events.on('openHistory', () => {
-			$showHistoryView = true;
-		})
+		)
 	);
-
-	onMount(() => {
-		return () => {
-			unsubscribeHistoryButton();
-		};
-	});
 </script>
