+++ apps/desktop/src/components/ReloadMenuAction.svelte
@@ -1,10 +1,8 @@
 <script lang="ts">
-	import { SHORTCUT_SERVICE } from '$lib/shortcuts/shortcutService.svelte';
+	import { SHORTCUT_SERVICE } from '$lib/shortcuts/shortcutService';
 	import { inject } from '@gitbutler/shared/context';
 
 	const shortcutService = inject(SHORTCUT_SERVICE);
 
-	shortcutService.on('reload', () => {
-		location.reload();
-	});
+	$effect(() => shortcutService.on('reload', () => location.reload()));
 </script>
