+++ apps/desktop/src/lib/updater/updater.test.ts
@@ -1,6 +1,7 @@
 import { EventContext } from '$lib/analytics/eventContext';
 import { PostHogWrapper } from '$lib/analytics/posthog';
 import { Tauri } from '$lib/backend/tauri';
+import { ShortcutService } from '$lib/shortcuts/shortcutService';
 import { getSettingsdServiceMock } from '$lib/testing/mockSettingsdService';
 import { UPDATE_INTERVAL_MS, UpdaterService } from '$lib/updater/updater';
 import { get } from 'svelte/store';
@@ -12,17 +13,17 @@ import type { Update } from '@tauri-apps/plugin-updater';
  * under the hood.
  */
 describe('Updater', () => {
-	let tauri: Tauri;
 	let updater: UpdaterService;
+	const tauri = new Tauri();
 	const MockSettingsService = getSettingsdServiceMock();
+	const shortcuts = new ShortcutService(tauri);
 	const settingsService = new MockSettingsService();
 	const eventContext = new EventContext();
 	const posthog = new PostHogWrapper(settingsService, eventContext);
 
 	beforeEach(() => {
 		vi.useFakeTimers();
-		tauri = new Tauri();
-		updater = new UpdaterService(tauri, posthog);
+		updater = new UpdaterService(tauri, posthog, shortcuts);
 		vi.spyOn(tauri, 'listen').mockReturnValue(async () => {});
 	});
